+++ packages/sdk/src/sdk/index.ts
@@ -2,5 +2,6 @@ export * from "./db";
 export * from "./notesManager";
 export * from "./merkleProver";
 export * from "./note";
+export * from "./asset";
 export * from "./signer";
 export * from "./utils";