+++ packages/sdk/src/sdk/notesManager/index.ts
@@ -1,12 +1,8 @@
 import { NotesDB } from "../db";
 import { IncludedNote, NoteTrait } from "../note";
 import { NocturneSigner } from "../signer";
-import {
-  decodeAsset,
-  Asset,
-  BaseJoinSplit,
-  EncryptedNote,
-} from "../../commonTypes";
+import { Asset, AssetTrait } from "../asset";
+import { BaseJoinSplit, EncryptedNote } from "../../commonTypes";
 
 export interface JoinSplitEvent {
   oldNoteANullifier: bigint;
@@ -50,10 +46,7 @@ export abstract class NotesManager {
     newJoinSplits: JoinSplitEvent[]
   ): Promise<void> {
     for (const e of newJoinSplits) {
-      const asset = decodeAsset(
-        e.joinSplit.encodedAsset.encodedAssetAddr,
-        e.joinSplit.encodedAsset.encodedAssetId
-      );
+      const asset = AssetTrait.decode(e.joinSplit.encodedAsset);
 
       await this.processEncryptedNote(
         e.joinSplit.newNoteACommitment,